chore: move CardGenerator to its own module
